+++ src/engine/graphic/GRMeter.cpp
@@ -44,7 +44,7 @@ GRMeter::GRMeter( const ARMeter * ar, GRStaff * curstaff, bool p_ownsAR )
 	mNeedsSpring = 1;
 	sconst = SCONST_METER;
 	fCurLSPACE = curstaff->getStaffLSPACE();
-	fNumericHeight	= fCurLSPACE * 0.85;		// adjust numeric glyph height
+	fNumericHeight	= fCurLSPACE * 0.85f;		// adjust numeric glyph height
 //	const TagParameterFloat* p = ar->getSize();
 //	mTagSize	   *= curstaff->getSizeRatio() * ar->getSize();
 //	mTagSize	   *= curstaff->getSizeRatio() * (p ? p->getValue() : 1.f);
@@ -149,7 +149,7 @@ string GRMeter::makeDenominatorString (const std::vector<Fraction>& meters) cons
 	if (getARMeter()->isSingleUnit())
 		strStream << meters[0].getDenominator();
 	else if (fGroupComplex) {
-		int previous = 0;
+		size_t previous = 0;
 		for (size_t i = 0; i < n; i++) {
 			int d = meters[i].getDenominator();
 
@@ -257,7 +257,7 @@ std::pair<float,float> GRMeter::GetXOffsets(VGDevice & hdc, const std::string& n
 {
 	NVRect nbb = computeBoundingBox (&hdc, num);
 	NVRect dbb = computeBoundingBox (&hdc, dnum);
-	pair<float, float> offsets = make_pair(0,0);
+	pair<float, float> offsets = make_pair(0.f,0.f);
 	float diff = (nbb.Width() - dbb.Width()) / 2;
 	if (diff > 0)	offsets.second = diff;
 	else			offsets.first = -diff;
@@ -290,7 +290,7 @@ float GRMeter::DrawNumericSingle(VGDevice & hdc, const string& num, const string
 	DrawSymbolStr (num.c_str(), xpos, y, hdc);
 
 	xpos = x + offsets.second;
-    y += (2 * fCurLSPACE + 1.7 * fCurLSPACE * (mTagSize - 1) );
+    y += (2 * fCurLSPACE + 1.7f * fCurLSPACE * (mTagSize - 1) );
 	DrawSymbolStr (dnum.c_str(), xpos, y, hdc);
 	return bb.Width();
 }
