+++ src/engine/graphic/GRTextHarmony.cpp
@@ -60,7 +60,7 @@ GRTextHarmony::GRTextHarmony(GRStaff * p_staff, const ARTextHarmony * ar)
 	if (myar)
 	{
 //		mFontSize = myar->getFSize(curLSPACE);
-		mFontSize = myar->getFSize() * curLSPACE / LSPACE;
+		mFontSize = int(myar->getFSize() * curLSPACE / LSPACE);
 		if (mFontSize == 0)
 			mFontSize = (int)(1.5f * LSPACE);
 		// we do not want that (do we?)
